refactor(arrow): extract common Arrow Native utilities into shared module
- Create arrow_common.rs module to eliminate code duplication
- Centralize parameter validation and output field creation
- Standardize error handling for Arrow operations
- Update all Arrow Native functions to use shared utilities
- Add documentation verification tools for consistency

+++ tests/performance/integration_benchmark.py
@@ -67,7 +67,7 @@ class IntegrationBenchmark:
 
         # ポートフォリオ統計
         results["total_value"] = np.sum(prices)
-        results["mean_price"] = np.mean(prices)
+        results["mean_price"] = float(np.mean(prices))
         results["portfolio_size"] = portfolio_size
         results["throughput"] = portfolio_size / results["valuation_time"]
